Make ADT interfaces self-contained
In order to accommodate the variety of Item types used in the exercises,
I've found it more convenient to make each ADT interface self-contained.
So, now each program needs only #include the interface, and the Item
definitions are loaded with it.

Depending on the case, the corresponding #include line must be
uncommented from the ADT. Currently, the include fails because there is
no generic Item.h present.